// web/src/api/generated/model/fiscalYearDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface FiscalYearDto {
  year: number;
  start: string;
  end: string;
  hasEnded: boolean;
  isClosed: boolean;
  profit: number;
  draftVouchers: number;
  /** @nullable */
  closingVoucherId: string | null;
}